+++ apps/web/app/mypage/posts/layout.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/screens/mypost/ui/MyPostLayout';
