+++ apps/web/src/screens/mypost/ui/MyPostLayout.tsx
@@ -1,9 +1,9 @@
 import SubNav from '@/shared/ui/layout/SubNav';
 
-export default function MyCommentLayout({ children }: { children: React.ReactNode }) {
+export default function MyPostLayout({ children }: { children: React.ReactNode }) {
     return (
         <>
-            <SubNav title='내가 작성한 댓글' />
+            <SubNav title='내가 작성한 게시글' />
             <main className='flex-1'>{children}</main>
         </>
     );
